[WIP] Only connect
