--- stories/ConnectedCounter/Controls/style.scss
@@ -1,18 +1,0 @@
-.counter .controls {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100px;
-
-    .control {
-        width: 30px;
-        height: 30px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        cursor: pointer;
-        line-height: 1;
-        border: 1px solid #d2d2d2;
-        box-shadow: #f4f4f4 1px 1px 1px 1px;
-    }
-}
