--- stories/ConnectedCounter/Display/style.scss
@@ -1,3 +1,0 @@
-.counter .display {
-    margin-bottom: 15px;
-}
