--- stories/ConnectedCounter/context.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const CounterContext = React.createContext();
-
-export default CounterContext;
