--- stories/ConnectedCounter/index.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import ReactFpContext from '../../src';
-import Display from './Display';
-import Controls from './Controls';
-import CounterContext from './context';
-import Strange from './Strange';
-import './style.scss';
-
-const ConnectedCounter = () => {
-    return (
-        <div className="counter">
-            <Controls/>
-            <Strange/>
-            <Display/>
-        </div>
-    );
-};
-
-export default ReactFpContext({
-    Context: CounterContext
-})(ConnectedCounter);
