--- stories/ConnectedCounter/style.scss
@@ -1,7 +1,0 @@
-.counter {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    justify-content: center;
-    user-select: none;
-}
