+++ stories/Counter/Display/index.js
@@ -1,18 +1,36 @@
-import React from 'react';
-import CounterContext from '../context';
+import React, { Profiler } from 'react';
+import _ from 'lodash';
+import { connect } from '../../../src';
 import './style.scss';
 
-const Display = () => {
-    const { context } = React.useContext(CounterContext);
-    const { count } = context;
+const callback = (id, phase, actualTime, baseTime, startTime, commitTime) => {
+    console.log(`${id}'s ${phase} phase:`);
+    console.log(`Actual time: ${actualTime}`);
+    console.log(`Base time: ${baseTime}`);
+    console.log(`Start time: ${startTime}`);
+    console.log(`Commit time: ${commitTime}`);
+};
 
-    console.log('Display updated');
-
+const Display = ({ count }) => {
     return (
-        <div className="display">
-            {count}
-        </div>
+        <Profiler id="DisplayPerformance" onRender={callback}>
+            <div className="display">
+                <table>
+                    {_.range(100).map((i) => (
+                        <tr key={i}>
+                            {_.range(100).map((j) => (
+                                <td key={j}>{count}</td>
+                            ))}
+                        </tr>
+                    ))}
+                </table>
+            </div>
+        </Profiler>
     )
 };
 
-export default Display;
+const useStateToProps = ({ context }) => ({
+    count: context.count
+});
+
+export default connect(useStateToProps)(Display);
