+++ stories/Counter/Strange/index.js
