+++ stories/Counter/index.js
@@ -3,17 +3,19 @@ import ReactFpContext from '../../src';
 import Display from './Display';
 import Controls from './Controls';
 import CounterContext from './context';
+import Strange from './Strange';
 import './style.scss';
 
-const Counter = () => {
+const ConnectedCounter = () => {
     return (
         <div className="counter">
-            <Display/>
             <Controls/>
+            <Strange/>
+            <Display/>
         </div>
     );
 };
 
 export default ReactFpContext({
     Context: CounterContext
-})(Counter);
+})(ConnectedCounter);
